Creating Clear Headings and Other Miscellaneous Changes
Created components folder and a template for the headings for the coding course pages.

Created index.css and moved the colour variables to it, as well as added all the fonts. Additionally, I imported the fonts to index.html.

Changed App.css and App.js to test and display the clear headings.

Also, imported react-icons for the information button icon.