Added Lesson Preview
Added a preview of the lesson when selecting a lesson from the learning path. The information it displays can be set in the PathPreviewInteraction file.

PathPreviewInteration was created to test the components together, it may possibly be removed in the future or reworked to have the two components in one file.

Still requires development, primarily in letting developer easily edit the information and to handle errors (there are other things, those are just the most notable).