Creating templates to add new courses/lessons/tests
Made sure that pages linked to each other (through routing) and that their information is different depending on what the user picks. The pages are just temporary with just the data displaying and will be updated/changed when the ui is developed for them.

Created a developer dashboard to add new courses/lessons/tests to make it easier to add, edit and delete content.

The data structure may need to change based on how we decide to store the data, but that should be easy to fix.

Still need to make sure that only developers can access this and possibly add toggles for when new pages are updated to let users see. Also needs error handling.