Attempted to make a dataFetching file
Attempted to make a file that will get the data for courses, lessons and tests, since that code is repeated in a lot of places and having a file dedicated to this is simpler.

Ran into issues, so reverted. But plan to restructure how data is fetched to make it more streamlined at a later date.