CSS Updates
Updated the UI on the home screen (logged out) and the sign in/log in pages.

Created a CodeOdysseyLogo asset that is the logo for the program. Also created Star asset, which will be used for decoration on the site to match planned UI.

Begun implementation on making sure the UI resizing correctly from 1920x1080 to 1024x600 to try an account for the average largest and smallest desktop sizes.