Fixed Hint System
Had an issue where the hint system would not show a prompt to set the timer nor would it show the hint after the elapsed time after implementing the function in the tests. Fixed both issues.

It still has some issues, but those need more testing to find out what's causing them.

Still need to fix up the CSS, but I am leaving that until the test display is done to make sure the UI matches.