Saving Changes
None of this code is done, but I am committing it now to save the changes I have made.

There are issues with deleting and editing data. Additionally, the dashboard on the developer dashboard is not complete. The code also needs to be cleaned up to make it easier to understand and to remove repeating chunks of code.

I have added a data saving and data deleting file to handle those functions to make the code more organised.