Update for Meeting
Updates are incomplete, but am pushing this so that I can show my work on my other device during group meeting.

Made it so that courses, lessons, and tests add, update, and delete data properly. Still has issues, but it is working a lot better.

Everything needs a CSS overhaul to make it easier to look at and use, but as of now it is functional. Everything also needs to be looked over to see if repeating code can be removed across files.