Implemented Footer UI
Added UI for the footer that resizes based on screen size. There is an issue with its location, as if a page does not have enough content on it the footer will just sit in the middle of the screen rather than the bottom.

Added footer to LoggedOutHomePage, ResourcesPage, and AboutUsPage to test it and its links.