Update Logged In Home Page UI
Updated the UI for the home page when the user is logged in. Made the UI resize for different screens. Still needs minor adjustments, but functionally works.

I merged the navigation bar CSS files since there is a lot of overlap. In the final merge the other navigation bar CSS files will need to be deleted.

Also added isAdmin to context. This will probably be changed, but I did this to test if the button to the developer dashboard is limited to just admins.