Updating CSS
Changed colors in (most) CSS files to use the variables in root.

Merged the navigation bar user and home into one file since its all basically the same code and it changes based on if user is authenticated or not. Also made it so its now above everything.

Fixed SubmitCode as it was missing some states and params after merging.

Updated the UI on the feedback system.