UI Updates
Updated the ProfilePicture component so that it can be reused in UserSettings. Still need to implement it.

Changed index.html so the tab for the website says “Code Odyssey” and has a little star icon

Changed the sign out button to redirect you to the home page and fixed the UI on the logged out home page

Refactored login and sign up pages since they are basically the exact same pages with minor changes. They’re now one file called AuthPage that exports two functions for login and sign up

Changed the user settings UI. All it needs is for the cropper CSS to be nicer, but I’ll do this after some bugs I found get fixed

Deleted various redundant files. I am keeping track of all of them in case they are added again.